Name and convention changed to Java 8
